+++ docs/CHANGELOG.md
@@ -3,7 +3,7 @@ Changelog
 
 1.6.2 (Unreleased)
 --------------------
-- Enh #495: Display info after delete event entry
+- Enh #495: Fix column `exdate` to delete more than 16 recurrence event entries
 
 1.6.1 (July 3, 2024)
 --------------------